+++ phy/cluster/manual/tests/test_session.py
@@ -20,7 +20,7 @@ from ....utils.dock import qt_app, _close_qt_after
 from ....utils.tempdir import TemporaryDirectory
 from ....utils.logging import set_level
 from ....io.mock import MockModel
-from ....io.mock.kwik import create_mock_kwik
+from ....io.kwik.mock import create_mock_kwik
 
 
 # Skip these tests in "make test-quick".
