+++ phy/plot/tests/test_features.py
@@ -12,9 +12,9 @@ from ...utils.logging import set_level
 from ..features import FeatureView
 from ...utils._color import _random_color
 from ...io.mock import (artificial_features,
-                                   artificial_masks,
-                                   artificial_spike_clusters,
-                                   artificial_spike_samples)
+                        artificial_masks,
+                        artificial_spike_clusters,
+                        artificial_spike_samples)
 from ...utils.testing import show_test
