+++ phy/plot/tests/test_traces.py
@@ -11,9 +11,9 @@ import numpy as np
 from ..traces import TraceView
 from ...utils._color import _random_color
 from ...io.mock import (artificial_traces,
-                                   artificial_masks,
-                                   artificial_spike_clusters,
-                                   )
+                        artificial_masks,
+                        artificial_spike_clusters,
+                        )
 from ...utils.testing import show_test
