+++ phy/plot/tests/test_waveforms.py
@@ -12,7 +12,7 @@ from ...utils.logging import set_level
 from ..waveforms import WaveformView
 from ...utils._color import _random_color
 from ...io.mock import (artificial_waveforms, artificial_masks,
-                                   artificial_spike_clusters)
+                        artificial_spike_clusters)
 from ...electrode.mea import staggered_positions
 from ...utils.testing import show_test
